+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - [#1578](https://github.com/influxdata/kapacitor/pull/1578): Add support for exposing logs via the API.
 - [#1605](https://github.com/influxdata/kapacitor/issues/1605): Add support for {{ .Duration }} on Alert Message property.
 - [#1644](https://github.com/influxdata/kapacitor/issues/1644): Add support for [JSON lines](https://en.wikipedia.org/wiki/JSON_Streaming#Line_delimited_JSON) for steaming HTTP logs.
+- [#1637](https://github.com/influxdata/kapacitor/issues/1637): Add new node Sideload, that allows loading data from files into the stream of data. Data can be loaded using a hierarchy.
 
 ### Bugfixes